reorganized file structure in /tests
